I created the MSBuild script to build the project.
I created the master MSBuild script to build and package the project for
deployment using NuGet. I created the
ImaginaryRealities.CommonLibrary.nuspec file to provide the properties
and list of files to be included in the NuGet package. I configured the
solution for NuGet package restore. I added a reference to the MSBuild
Extensions project to the solution.