I updated the project to build correctly.
I renamed the SemanticVersionNumberAttribute class to
SemanticVersionAttribute. I had to do this because I previously released
the MSBuild Extensions project that depended on that name. I updated the
CommonLibrary.msbuild script to generate the VersionInfo.cs file for
each build configuration. I renamed ImaginaryRealities.Common.dll to
ImaginaryRealities.Framework.dll. I did this because the MSBuild
Extensions used that namespace.